requirements.txt
pytest==8.1.1
allure-pytest==2.13.5
kubernetes==29.0.0
PyYAML==6.0.1